PythonProject/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_login
flask_wtf
wtforms
psycopg2-binary
werkzeug
